add InteractiveFunc param type int8

+++ service/varArvs.go
@@ -26,7 +26,7 @@ type (
 	SetWinScoreFunc           func(string, RoundUploadStruct, int64) error
 	LotteryFunc               func(string, string, int64) map[string]int64
 	WebsocketFunc             func(msg *pmsg.MessageBody) error
-	InteractiveFunc           func(roomId, roundId string, label int) bool //自动选边
+	InteractiveFunc           func(roomId, roundId string, label int8) bool //自动选边
 	GiftExtendInfoFunc        func() string
 	ScrollFunc                func(*string)
 	InitFunc                  func(bool)
